Fix arg types for sound and window

+++ platforms/Mac OS/plugins/SoundPlugin/sqMacUnixInterfaceSound.c
@@ -39,14 +39,14 @@ sqInt snd_AvailableSpace(void)
   return sound_AvailableSpace();
 }
 
-sqInt snd_InsertSamplesFromLeadTime(sqInt frameCount, void *srcBufPtr, sqInt samplesOfLeadTime)
+sqInt snd_InsertSamplesFromLeadTime(sqInt frameCount, void * srcBufPtr, sqInt samplesOfLeadTime)
 {
-  return sound_InsertSamplesFromLeadTime(frameCount, srcBufPtr, samplesOfLeadTime);
+  return sound_InsertSamplesFromLeadTime(frameCount, (int)srcBufPtr, samplesOfLeadTime);
 }
 
 sqInt snd_PlaySamplesFromAtLength(sqInt frameCount, void *srcBufPtr, sqInt startIndex)
 {
-  return sound_PlaySamplesFromAtLength(frameCount, srcBufPtr, startIndex);
+  return sound_PlaySamplesFromAtLength(frameCount, (int) srcBufPtr, startIndex);
 }
 
 sqInt snd_PlaySilence(void)
@@ -81,9 +81,9 @@ double snd_GetRecordingSampleRate(void)
   return sound_GetRecordingSampleRate();
 }
 
-sqInt snd_RecordSamplesIntoAtLength(void *buf, sqInt startSliceIndex, sqInt bufferSizeInBytes)
+sqInt snd_RecordSamplesIntoAtLength(void * buf, sqInt startSliceIndex, sqInt bufferSizeInBytes)
 {
-  return sound_RecordSamplesIntoAtLength(buf, startSliceIndex, bufferSizeInBytes);
+  return sound_RecordSamplesIntoAtLength((int)buf, startSliceIndex, bufferSizeInBytes);
 }
 
 /* Acoustic echo-cancellation (AEC) is not supported on Macintosh yet. */
